% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OutSamples.R
\name{OutSamples}
\alias{OutSamples}
\title{Posterior Samples of the Outcomes}
\usage{
OutSamples(obj1, obj0, q)
}
\arguments{
\item{obj1}{The fitted model of the observed data under Z=1 from DPdensity}

\item{obj0}{The fitted model of the observed data under Z=0 from Dpdensity}

\item{q}{A dimension of the observed data, i.e., number of covariates plus 2}
}
\value{
Y1 Posterior samples of E[Y1]

Y0 Posterior samples of E[Y0]

TMean The posterior mean of the total effect

TCI 95\% C.I. of the total effect
}
\description{
Obtain posterior sampels of the outcomes, E[Y1] and E[Y0].
}

